backend/requirements.txt
# FastAPI and Web Server
fastapi==0.110.0
uvicorn==0.27.1
starlette==0.36.3
jiter==0.8.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
websockets==12.0
python-multipart==0.0.9
orjson==3.10.15
sniffio==1.3.1
anyio==4.8.0

# Database
SQLAlchemy==1.4.54  # Note: Using 1.4.x for compatibility
alembic==1.12.1
asyncpg==0.29.0
databases==0.8.0
greenlet==3.1.1  # Required by SQLAlchemy for async
psycopg2-binary==2.9.9  # PostgreSQL driver

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
ecdsa==0.19.0
rsa==4.9
pyasn1==0.6.1
email_validator==2.2.0
PyJWT==2.10.1
dnspython==2.7.0

# AI & LLM Integration
openai==1.65.4
langchain==0.3.20
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
tiktoken==0.9.0  # For token counting
tenacity==9.0.0  # For retrying API calls

# Vector Database & Embeddings
faiss-cpu==1.10.0  # Vector database for similarity search
sentence-transformers==3.4.1  # For creating embeddings
transformers==4.49.0
tokenizers==0.21.0
torch==2.6.0
huggingface-hub==0.29.3
scikit-learn==1.6.1  # Used by some embedding models
scipy==1.15.2
numpy==2.2.3
pandas==2.2.3

# Caching
redis==5.2.1  # For Redis-based caching

# Image Processing
pytesseract==0.3.13  # OCR for document scanning
pillow==11.1.0  # Image processing

# Supabase Integration
supabase==2.0.3
supafunc==0.3.3
gotrue==1.3.1
postgrest==0.13.2
realtime==1.0.6
storage3==0.6.1

# Utilities
python-dotenv==1.0.1  # Environment variable management
Jinja2==3.1.6  # Template engine
MarkupSafe==3.0.2
pydantic==2.10.6  # Data validation
pydantic_core==2.27.2
annotated-types==0.7.0
typing_extensions==4.12.2
StrEnum==0.4.15
certifi==2025.1.31
idna==3.10
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.3.0
click==8.1.8
Mako==1.3.9
python-dateutil==2.9.0.post0
six==1.17.0
pytz==2025.1
tzdata==2025.1
distro==1.9.0
filelock==3.17.0
fsspec==2025.3.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
mpmath==1.3.0
networkx==3.4.2
packaging==24.2
regex==2024.11.6
safetensors==0.5.3
setuptools==76.0.0
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.1
zstandard==0.23.0
PyYAML==6.0.2
deprecation==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
iniconfig==2.0.0
pluggy==1.5.0
